fix: add Vim to index
Currently the website only updates if the title of the new markdown file is added to index.txt in addition to being created and saved to the static directory. This can be simplified by creating a server that gets the contents of the static directory and generates links accordingly. Trying with node.